Create AbpODataDtoController
This is a simple mod, utilising Automapper, to create a contoller base that enables the Use of DTO's with odata, Rather than the raw Entities.
Using a DTO has many advantages, as we know, for code security, but also, it keeps the entitys clean of OData EDM conventions annotations such as [DataMember], [IgnoreDataMember] etc. It allows us to flatten nested joins, created caluclated columns etc.

This implementation allows the user to define all of the EDM Annotations in the DTO...
It just requires the reuglar automapper dto mapping and ODataConventionModelBuilder with: .EntitySet<EntityDTO>(...)...

I think some further OData development like this, could realy make ABP a much more flexable product.